backend/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Backend for the LangGraph agent"
authors = [
    { name = "Dave Boku", email = "davesah.bs@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    # Core dependencies
    "langgraph>=0.2.6",
    "langchain>=0.3.19",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.13",
    "langchain-core>=0.1.17",
    "langgraph-sdk>=0.1.57",
    "langgraph-cli>=0.1.71",
    "langgraph-api>=0.2.51",
    "langchain-google-genai",
    "google-genai",
    
    # Database and Vector Store
    "supabase>=2.3.0",
    "python-dotenv>=1.0.1",
    
    # Web Framework
    "fastapi>=0.109.0",
    "starlette>=0.36.0",
    "uvicorn>=0.27.0",
    
    # Data Models and Validation
    "pydantic>=2.6.0",
    "typing-extensions>=4.9.0",
    
    # Optional but recommended
    "python-multipart>=0.0.9",
    "httpx>=0.26.0",
]


[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
    "pytest>=8.3.5",
    "langgraph-cli[inmem]>=0.1.71",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.1.71",
    "pytest>=8.3.5",
]
